requirements.txt
amqp==2.5.2
asgiref==3.2.5
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-redis-cache==2.1.0
gunicorn==20.0.4
importlib-metadata==1.5.1
joblib==0.14.1
kombu==4.6.8
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
scikit-learn==0.22.2.post1
scikit-surprise==1.1.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
vine==1.3.0
whitenoise==5.0.1
wincertstore==0.2
zipp==3.1.0
